+++ src/commands/tenant/editAccessRequestConfigCommand.ts
@@ -1,14 +1,14 @@
-import { TenantTreeItem } from "../models/IdentityNowTreeItem";
+import { TenantTreeItem } from "../../models/IdentityNowTreeItem";
 import * as vscode from 'vscode';
-import { getResourceUri } from "../utils/UriUtils";
-import { openPreview } from "../utils/vsCodeHelpers";
+import { getResourceUri } from "../../utils/UriUtils";
+import { openPreview } from "../../utils/vsCodeHelpers";
 
 export class EditAccessRequestConfigCommand {
 
     async execute(node: TenantTreeItem): Promise<void> {
         console.log("> EditAccessRequestConfigCommand.execute", node);
 
-        const publicIdentitiesConfigUri = getResourceUri(node.tenantName,
+        const accessRequestConfigUri = getResourceUri(node.tenantName,
             'access-request-config',
             null,
             "Access Request Configuration");
@@ -18,7 +18,7 @@ export class EditAccessRequestConfigCommand {
             title: 'Opening Access Request Configuration...',
             cancellable: false
         }, async (task, token) => {
-            await openPreview(publicIdentitiesConfigUri)
+            await openPreview(accessRequestConfigUri)
         });
     }
 }